requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
bidict==0.23.1
bleak==0.22.3
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.2
coverage==7.8.0
cycler==0.12.1
Django==4.1.12
django-admin-datta==1.0.10
django-api-generator==1.0.17
django-debug-toolbar==4.2.0
django-dynamic-datatb==1.0.25
djangorestframework==3.15.1
docutils==0.21.2
exceptiongroup==1.2.2
fastapi==0.115.12
fonttools==4.57.0
frozenlist==1.6.0
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
itsdangerous==2.2.0
jinja2==3.1.6
kiwisolver==1.4.8
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.1
multidict==6.4.3
nicegui==2.15.0
numpy==2.2.5
orjson==3.10.16
packaging==25.0
pandas==2.2.3
pillow==11.2.1
propcache==0.3.1
pscript==0.7.7
pygments==2.19.1
pyparsing==3.2.3
pydantic==2.11.3
pydantic-core==2.33.1
PyYAML==6.0.2
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.20
python-socketio==5.13.0
pytz==2025.2
requests==2.32.3
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
starlette==0.46.2
tabulate==0.9.0
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
vbuild==0.8.2
watchfiles==1.0.5
websockets==15.0.1
whitenoise==6.5.0
wsproto==1.2.0
yarl==1.20.0
